services/ingestion-agent-service/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and serialization
pydantic==2.5.0
pydantic[email]==2.5.0

# Environment variables
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# File uploads and processing
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
Pillow==10.1.0

# AI and LLM
openai>=1.6.1
langchain==0.1.0
langchain-openai==0.0.2

# WebSocket for real-time chat
websockets==12.0

# HTTP requests
requests==2.31.0
httpx==0.25.2

# Email validation
email-validator==2.1.0

# Security
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Code formatting and linting
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Utilities
python-dateutil==2.8.2
